requirements.txt
streamlit==1.56.0
pandas==2.3.3
pdfplumber==0.11.9
PyMuPDF==1.27.2.2
opencv-python==4.13.0.92
paddleocr==3.3.3
paddlepaddle==3.2.0
llama_cpp_python==0.2.90
google-genai
